backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
pytest==6.2.4
pytest-django
django-cors-headers
django-crispy-forms
django-imagekit
crispy_bootstrap5